segmented line: commit to this repo, sinceit has nowhere else to live
Adds a new chart type, "SegmentedLine", to Chartist.  Pretty
inflexible and unstandardized, *but* it's kind of cool: allows you to
break lines up into usb-series based on a criterion & a function that
applies the criterion to the series.

Lots of improvements possible:

- allow segmentation by series if desired (and if a given series is
unsegmented, just run it asa line instead -- maybe just return
Chartist.Line(series), or something?  probably not quite so easy.
- make structure of segments conform more perfectly to that of series,
so that all Chartist series methods willwork on them.
- check to see if I'm misusing meta horribly.
- try to remove ad hoc default vlaues for variables.